Remove pip/setuptools/wheel from requirements.txt
`pip`, `setuptools`, and `wheel` generally aren't specified in a `requirements.txt`.  The first two won't be particularly useful since they wouldn't be installed prior to the installation of the other requirements.  Also note that on Windows you have to `python -m pip` instead of just `pip` to be able to upgrade `pip`.